% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dat.R
\name{unescape_character_reference}
\alias{unescape_character_reference}
\title{Unescape character references}
\usage{
unescape_character_reference(str)
}
\arguments{
\item{str}{a string vector.}
}
\value{
a string vector.
}
\description{
Unescape character references
}
